# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API DECEA",
        "version": "1.0"
    },
    "paths": {
...
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Sign up",
                "description": "Register",
                "requestBody": {
                    "description": "Values for register",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "name",
                                    "document",
                                    "social_name",
                                    "alias_name",
                                    "document_company",
                                    "document_company_secondary"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "user1"
                                    },
                                    "document": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "03585965236"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "exemplo2@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123456"
                                    },
                                    "social_name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Fundacao STDP LTDA"
                                    },
                                    "alias_name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Fundacao STDP"
                                    },
                                    "document_company": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "63565720000125"
                                    },
                                    "document_company_secondary": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "5066485-8"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/{company}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Users",
                "description": "Retorna uma lista de usuarios",
                "operationId": "e1f11320cd661e4d9523004c0665e104",
                "parameters": [
                    {
                        "name": "company",
                        "in": "path",
                        "description": "ID of company",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "2"
                                        },
                                        "company_id": {
                                            "type": "string",
                                            "example": "2"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "user1"
                                        },
                                        "document": {
                                            "type": "string",
                                            "example": "035.859.652-36"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "exemplo2@gmail.com"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}